+++ src/MainWindow.h
@@ -7,6 +7,7 @@
 #include <QtWidgets/QErrorMessage>
 #include <QtWidgets/QProgressDialog>
 #include <QDropEvent>
+#include <osgViewer/Viewer>
 
 #include "ui_MainWindow.h"
 #include "OSGViewerWidget.h"
@@ -17,13 +18,16 @@
 #include "labelCanvasView.h"
 #include "resources/ERDisplay.h"
 
-//#include "VSimApp.h"
 
+//#include "VSimApp.h"
 //extern osgViewer::Viewer* g_viewer;
 
 class VSimApp;
+class OSGViewerWidget;
 class labelCanvas;
 class labelCanvasView;
+class ModelOutliner;
+class TimeSlider;
 
 class MainWindow : public QMainWindow
 {
@@ -38,8 +42,13 @@ public:
 	void LoadingDialog(const std::string& msg);
 
 	// linking
-	//OSGViewerWidget *getViewerWidget() const { return m_osg_widget;	}
-	osgViewer::Viewer* getViewer() const { return m_osg_widget->getViewer(); };
+	OSGViewerWidget *getViewerWidget() const;
+	osgViewer::Viewer* getViewer() const;
+
+	MainWindowTopBar *topBar() const;
+	labelCanvasView *canvasView() const;
+	ModelOutliner *outliner() const;
+	TimeSlider *timeSlider() const;
 
 	// event stuff
 	void paintEvent(QPaintEvent* event);
@@ -76,18 +85,27 @@ signals:
 	void sSaveCurrent();
 	void sNew();
 	void sImportModel(const std::string&);
+	void sImportNarratives();
+	void sExportNarratives();
+
+	void sDebugOSG();
+	void sDebugCamera();
+	void sDebugControl();
 
 	// ui, signal emitters
 public:
-	Ui::MainWindow ui;
+	std::unique_ptr<Ui::MainWindow> ui;
 
 public:
+	QUndoStack *m_undo_stack;
+
 	OSGViewerWidget *m_osg_widget;
 	labelCanvas *m_drag_area;
-	QUndoStack *m_undo_stack;
 	labelCanvasView *m_view;
 	ERDisplay *m_display;
 	osg::ref_ptr<EResourceGroup> m_resource_group;
+	ModelOutliner *m_outliner;
+	TimeSlider *m_time_slider;
 };
 
 enum SelectionCommandWhen {