+++ src/VSimApp.cpp
@@ -18,29 +18,60 @@
 #include "Util.h"
 #include "deprecated/narrative/Narrative.h"
 #include "narrative/NarrativeGroup.h"
+#include "narrative/NarrativeControl.h"
+#include "OSGViewerWidget.h"
+#include "MainWindow.h"
+#include "ModelOutliner.h"
+#include "TimeSlider.h"
 
 #define OPTIMIZE 0
 
 
 VSimApp::VSimApp(MainWindow* window)
-	: m_window(window)
+	: m_window(window),
+	m_filename(""),
+	m_root(new VSimRoot),
+	m_model_table_model(m_root->models())
 {
 	m_viewer = window->getViewer();
+
+	// timers
+	m_timer = new QTimer;
+	m_timer->setInterval(10);
+	m_timer->setSingleShot(false);
+	m_dt_timer = new QElapsedTimer;
+	m_timer->start();
+	m_dt_timer->start();
+	connect(m_timer, &QTimer::timeout, this, &VSimApp::updateTime);
+
 	m_narrative_control = new NarrativeControl(this, m_window);
-	m_narrative_player = new NarrativePlayer(this, m_window, m_narrative_control);
+
+	// Narrative player
+	m_narrative_player = new NarrativePlayer(this, m_narrative_control);
+	connect(window->topBar()->ui.play_2, &QPushButton::pressed, m_narrative_player, &NarrativePlayer::play);
+	connect(window->topBar()->ui.pause_2, &QPushButton::pressed, m_narrative_player, &NarrativePlayer::stop);
+	connect(m_narrative_player, &NarrativePlayer::updateCamera, m_window->getViewerWidget(), &OSGViewerWidget::setCameraMatrix);
+	connect(this, &VSimApp::tick, m_narrative_player, &NarrativePlayer::update);
+	//connect(this, &VSimApp::foo, window->getViewerWidget(), static_cast<void(OSGViewerWidget::*)()>(&OSGViewerWidget::update));
+	connect(this, &VSimApp::tick, window->getViewerWidget(), static_cast<void(OSGViewerWidget::*)()>(&OSGViewerWidget::update));
+
+	// This is a really awkward place... but this has to be done after setting model
+	m_window->outliner()->setModel(&m_model_table_model);
+	m_window->outliner()->header()->resizeSection(0, 200);
+	m_window->outliner()->resize(505, 600);
 
 	connect(window, &MainWindow::sOpenFile, this, &VSimApp::openVSim);
 	connect(window, &MainWindow::sSaveFile, this, &VSimApp::saveVSim);
 	connect(window, &MainWindow::sImportModel, this, &VSimApp::importModel);
-	connect(window, &MainWindow::sNew, this, &VSimApp::reset);
+	connect(window, &MainWindow::sNew, this, &VSimApp::init);
 	connect(window, &MainWindow::sSaveCurrent, this, &VSimApp::saveCurrentVSim);
-	connect(window->ui.actionImport_Narratives, &QAction::triggered, this, &VSimApp::importNarratives);
-	connect(window->ui.actionExport_Narratives, &QAction::triggered, this, &VSimApp::exportNarratives);
+	connect(window, &MainWindow::sImportNarratives, this, &VSimApp::importNarratives);
+	connect(window, &MainWindow::sExportNarratives, this, &VSimApp::exportNarratives);
 
-	connect(m_window->ui.actionOSG_Debug, &QAction::triggered, this, [this]() {m_root->debug(); });
-	connect(m_window->ui.actionCamera_Debug, &QAction::triggered, this, &VSimApp::debugCamera);
+	connect(window, &MainWindow::sDebugOSG, m_root, &VSimRoot::debug);
+	connect(window, &MainWindow::sDebugCamera, this, &VSimApp::debugCamera);
 
-	reset();
+	initWithVSim(m_root);
 }
 
 bool VSimApp::init()
@@ -49,13 +80,7 @@ bool VSimApp::init()
 	initWithVSim(new VSimRoot);
 	return true;
 }
-bool VSimApp::initWithModel(osg::Node *model)
-{
-	init();
-	m_root->models()->addChild(model);
-	m_window->m_osg_widget->reset();
-	return true;
-}
+
 bool VSimApp::initWithVSim(osg::Node *new_node)
 {
 	VSimRoot *root = dynamic_cast<VSimRoot*>(new_node);
@@ -64,27 +89,36 @@ bool VSimApp::initWithVSim(osg::Node *new_node)
 		root = new VSimRoot(new_node->asGroup());
 	}
 	else {
-		qDebug() << "is in fact a vsimroot";
+		qDebug() << "Root is a VSimRoot";
 	}
-	m_root = root;
-	m_viewer->setSceneData(m_root->models()); // ideally this would be only models, but its easy to mess things up
-
-	m_narrative_control->load(m_root->narratives());
-
+	
+	// move all of the gui stuff over to the new root
+	m_viewer->setSceneData(root->models()); // ideally this would be only models, but its easy to mess things up
+	m_model_table_model.setGroup(root->models());
+	m_narrative_control->load(root->narratives());
+	m_window->timeSlider()->setGroup(root->models());
+	
 	m_window->m_undo_stack->clear();
 	m_window->m_osg_widget->reset();
-	//m_viewer->getCamera()->setProjectionMatrixAsPerspective(75.0f, )
+
+	// dereference the old root, apply the new one
+	m_root = root;
 	
 	return true;
 }
-void VSimApp::reset()
+
+
+void VSimApp::addModel(osg::Node *node, const std::string &name)
 {
-	init();
+	m_root->models()->addChild(node);
+	node->setName(name);
+	m_window->m_osg_widget->reset();
 }
 
 bool VSimApp::importModel(const std::string& filename)
 {
 	osg::ref_ptr<osg::Node> loadedModel(NULL);
+	// TODO: special import for vsim files
 
 	// otherwise
 	loadedModel = osgDB::readNodeFile(filename);
@@ -93,8 +127,7 @@ bool VSimApp::importModel(const std::string& filename)
 		QMessageBox::warning(m_window, "Import Error", "Error loading file " + QString::fromStdString(filename));
 		return false;
 	}
-	m_root->models()->addChild(loadedModel);
-	m_window->m_osg_widget->reset();
+	addModel(loadedModel, Util::getFilename(filename));
 	return true;
 }
 
@@ -146,7 +179,13 @@ bool VSimApp::openVSim(const std::string & filename)
 	else {
 		qDebug() << "loading a non osg model";
 		loadedModel = osgDB::readNodeFile(filename);
-		init_success = initWithModel(loadedModel);
+		if (!loadedModel) {
+			QMessageBox::warning(m_window, "Load Error", "Failed to load model " + QString::fromStdString(filename));
+			return false;
+		}
+		init();
+		addModel(loadedModel, Util::getFilename(filename));
+		init_success = true;
 	}
 
 	if (!init_success) {
@@ -221,14 +260,14 @@ bool VSimApp::exportNarratives()
 	std::set<int> selection;
 	if (level == NarrativeControl::NARRATIVES) {
 		// export the entire selection
-		selection = m_window->ui.topBar->ui.narratives->getSelection();
+		selection = m_window->topBar()->ui.narratives->getSelection();
 		if (selection.empty()) {
 			QMessageBox::warning(m_window, "Export Narratives Error", "No narratives selected");
 			return false;
 		}
 	}
 	else {
-		int nar = m_narrative_control->getCurrentNarrative();
+		int nar = m_narrative_control->getCurrentNarrativeIndex();
 		if (nar == -1) return false;
 		selection.insert(nar);
 	}
@@ -360,4 +399,11 @@ void VSimApp::debugCamera()
 	Util::quatToYPR(rot, &y, &p, &r);
 	std::cout << "matrix " << matrix << "\ntranslation " << trans << "\nscale " << scale << "\nrotation " << rot << "\n";
 	qInfo() << "ypr" << y * 180 / M_PI << p * 180 / M_PI << r * 180 / M_PI;
+}
+
+void VSimApp::updateTime()
+{
+	double dt = m_dt_timer->nsecsElapsed() / 1.0e9;
+	m_dt_timer->restart();
+	emit tick(dt);
 }